refactor(backend): modularize CompositionRoot
