+++ packages/backend/src/modules/notifications/NotificationsModule.ts
@@ -1,0 +1,9 @@
+import { TransactionalEmailAPI } from './TransactionalEmailAPI';
+
+export class NotificationsModule {
+  private transactionalEmailAPI = new TransactionalEmailAPI();
+
+  getTransactionalEmailAPI() {
+    return this.transactionalEmailAPI;
+  }
+}
