+++ packages/backend/src/shared/CompositionRoot.ts
@@ -1,17 +1,9 @@
-import { Database } from './database/Database';
-import { ContactListAPI } from '../modules/marketing/ContactListAPI';
-import { TransactionalEmailAPI } from '../modules/notifications/TransactionalEmailAPI';
+import { SharedModule } from './SharedModule';
+import { UsersModule } from '../modules/users/UsersModule';
+import { NotificationsModule } from '../modules/notifications/NotificationsModule';
+import { PostsModule } from '../modules/posts/PostsModule';
+import { MarketingModule } from '../modules/marketing/MarketingModule';
 import { errorHandler } from './errorHandler';
-import { prisma } from './database/prisma/prisma';
-import { WebServer } from './WebServer';
-import { PostsService } from '../modules/posts/PostsService';
-import { UsersService } from '../modules/users/UsersService';
-import { UsersRepository } from '../modules/users/UsersRepository';
-import { MarketingController } from '../modules/marketing/MarketingController';
-import { MarketingService } from '../modules/marketing/MarketingService';
-import { PostsController } from '../modules/posts/PostsController';
-import { PostsRepository } from '../modules/posts/PostsRepository';
-import { UsersController } from '../modules/users/UsersController';
 
 export class CompositionRoot {
   private static Instance: CompositionRoot;
@@ -22,64 +14,27 @@ export class CompositionRoot {
     return this.Instance;
   }
 
-  // Database
-  private database: Database;
-
-  // Repositories
-  private usersRepository: UsersRepository;
-  private postsRepository: PostsRepository;
-
-  // External services
-  private transactionalEmailAPI: TransactionalEmailAPI;
-  private contactListAPI: ContactListAPI;
-
-  // Services
-  private usersService: UsersService;
-  private postsService: PostsService;
-  private marketingService: MarketingService;
-
-  // Controllers
-  private usersController: UsersController;
-  private postsController: PostsController;
-  private marketingController: MarketingController;
-
-  // Web server
-  private webServer: WebServer;
+  private sharedModule: SharedModule;
+  private notificationsModule: NotificationsModule;
+  private usersModule: UsersModule;
+  private postsModule: PostsModule;
+  private marketingModule: MarketingModule;
 
   private constructor() {
-    this.database = new Database(prisma);
+    this.sharedModule = new SharedModule();
+    const webServer = this.sharedModule.getWebServer();
+    const database = this.sharedModule.getDatabase();
 
-    this.usersRepository = new UsersRepository(this.database);
-    this.postsRepository = new PostsRepository(this.database);
+    this.notificationsModule = new NotificationsModule();
+    this.usersModule = new UsersModule(database, this.notificationsModule.getTransactionalEmailAPI(), webServer);
+    this.postsModule = new PostsModule(database, webServer);
+    this.marketingModule = new MarketingModule(webServer);
 
-    this.transactionalEmailAPI = new TransactionalEmailAPI();
-    this.contactListAPI = new ContactListAPI();
-
-    this.usersService = new UsersService(this.usersRepository, this.transactionalEmailAPI);
-    this.postsService = new PostsService(this.postsRepository);
-    this.marketingService = new MarketingService(this.contactListAPI);
-
-    this.usersController = new UsersController(this.usersService);
-    this.postsController = new PostsController(this.postsService);
-    this.marketingController = new MarketingController(this.marketingService);
-
-    this.webServer = this.createWebServer();
-  }
-
-  private createWebServer() {
-    const webServer = new WebServer({
-      port: Number(process.env.PORT) || 3000,
-    });
-
-    webServer.registerController('/users', this.usersController);
-    webServer.registerController('/posts', this.postsController);
-    webServer.registerController('/marketing', this.marketingController);
+    // registering error handling middleware at last
     webServer.registerMiddleware(errorHandler);
-
-    return webServer;
   }
 
   getWebServer() {
-    return this.webServer;
+    return this.sharedModule.getWebServer();
   }
 }
